chore: address pr comments, improve responses, log tx hashes
